big, fat, and horrible commit (Global formulation)
- revives a global integrator (MoreauJeanGOSI)
- revives the global formulation for FC (GlobalFrictionContact)
- seems to be working on BouncingBallNEGlobalTS
- introduction of new functions to fill a csc matrix

Note that there is too much in common between MoreauJeanGOSI and
MoreauJeanOSI to keep them both separated. After review and feedback, I
would like to merge the 2. The only major change is that the WMap is
gone in MoreauJeanGOSI. Also we need to separate the contribution of
the NSL from the computation of the smooth part. Then it should be easy
to share almost all the code. The only optimization that could be done
would be to skip the computation of the new velocity, since it is
readily available.