[mechanics proposed] give mechanics_io.py a 'backend' selector, select backend per example
For now this allows to switch between the two Mechanics APIs
(implementations "proposed" and "original") by setting a single flag,
but in the future could also be used to switch from Bullet to another
collision backend.

Tested all the examples, and selected the original API for ones that
did not work correctly: there is still a missing Cylinder object, and
there appears to be a contact bug causing objects to fall through the
plane.